fixed the map bug

+++ ElEngino/includes/SDLGraphics.h
@@ -30,7 +30,7 @@ namespace engino {
 		virtual size_t LoadTexture(const char* filename) override;
 		virtual size_t LoadFont(const char* filename, int fileSize) override;
 
-		virtual void DrawSprite(int x, int y, int w, int h, int angle, size_t fileId) override;
+		virtual void DrawSprite(int x, int y, int w, int h, double angle, size_t fileId) override;
 		virtual void DrawSprite(const RectI& src, const RectF& dst,
 			double angle, const Flip& flip, const Color& color, size_t fileId) override;
 		virtual void DrawSprite(const RectF& dst, const Color& color, size_t fileId) override;
